# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloud-Ready School Management ERP API - Phase 1",
        "description": "Complete RESTful API specifications for Multi-Campus School Management ERP. Encompasses Institute Management, Multi-Campus Branches, Academic Years & Terms, Spatie Role-Based Access Control (RBAC), Global Settings, Classes, Sections, Streams, Houses, Subjects, Employee & Staff Management, and Student Admissions & Profiles.",
        "version": "1.0.0",
        "contact": {
            "name": "School ERP Support",
            "email": "support@schoolerp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Laravel Development API Server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Sanctum API Bearer Token"
            }
        }
    },
    "paths": {
...
        "/dashboard/kpi": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get Phase 1 Dashboard KPIs",
                "description": "Returns total students, faculty, classes, active session, branch counts and recent admissions.",
                "responses": {
                    "200": {
                        "description": "KPI metrics returned"
                    }
                }
            }
        },
...
        "/classes": {
            "get": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Classes with Sections",
                "responses": {
                    "200": {
                        "description": "List of classes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Class / Grade",
                "responses": {
                    "201": {
                        "description": "Class created"
                    }
                }
            }
        },
        "/classes/{id}": {
            "get": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get Class Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Class",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Class",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class deleted"
                    }
                }
            }
        },
        "/classes/{classId}/sections": {
            "post": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Add Section to Class",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Section created"
                    }
                }
            }
        },
        "/sections/{sectionId}": {
            "put": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Section",
                "parameters": [
                    {
                        "name": "sectionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Section",
                "parameters": [
                    {
                        "name": "sectionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section deleted"
                    }
                }
            }
        },
        "/streams": {
            "get": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Streams (Science, Commerce, Arts)",
                "responses": {
                    "200": {
                        "description": "List of streams"
                    }
                }
            },
            "post": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Stream",
                "responses": {
                    "201": {
                        "description": "Stream created"
                    }
                }
            }
        },
        "/houses": {
            "get": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Houses / Groups",
                "responses": {
                    "200": {
                        "description": "List of houses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Class & Section Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create House",
                "responses": {
                    "201": {
                        "description": "House created"
                    }
                }
            }
        },
...
    }
}